Flake8 fixes
