+++ webapp/wsgi.py
@@ -10,7 +10,7 @@ sys.path.append(BASE_DIR)
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "webapp.settings")
 
-from django.core.wsgi import get_wsgi_application
-from dj_static import Cling
+from django.core.wsgi import get_wsgi_application  # noqa: E402
+from dj_static import Cling                        # noqa: E402
 
 application = Cling(get_wsgi_application())
